Split investments models into their own files
