+++ accounts/tests/factories/users.py
@@ -1,6 +1,7 @@
 from django.contrib.auth.models import User
 import factory
 
+
 class UserFactory(factory.django.DjangoModelFactory):
     class Meta:
         model = User
