+++ investments/models/__init__.py
@@ -1,0 +1,2 @@
+from .fund import Fund
+from .subscription import Subscription
