+++ investments/models/fund.py
@@ -1,0 +1,13 @@
+from django.db import models
+from django.utils.translation import gettext_lazy as _
+
+
+class Fund(models.Model):
+    name = models.CharField(max_length=100, unique=True)
+
+    def __str__(self):
+        return self.name
+
+    class Meta:
+        verbose_name = _('fund')
+        verbose_name_plural = _('funds')
