+++ investments/models/subscription.py
@@ -1,16 +1,8 @@
 from django.db import models
 from django.utils.translation import gettext_lazy as _
+
 from accounts.models import Profile
-
-class Fund(models.Model):
-    name = models.CharField(max_length=100, unique=True)
-
-    def __str__(self):
-        return self.name
-
-    class Meta:
-        verbose_name = _('fund')
-        verbose_name_plural = _('funds')
+from investments.models import Fund
 
 class Subscription(models.Model):
     amount = models.PositiveIntegerField()
